pyproject.toml
[project]
name = "mocli"
version = "1.0.0"
description = "AI-Powered Desktop Companion — 一个运行在 Windows 桌面的 AI 伴侣助手，具备视觉指向、语音唤醒、语音识别和语音合成能力。"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "keyboard==0.13.5",
    "markdown>=3.10.2",
    "numpy>=1.26.0",
    "onnxruntime-gpu>=1.24.4",
    "pillow>=12.0.0",
    "pyaudio==0.2.14",
    "pypinyin>=0.55.0",
    "pyqt6==6.11.0",
    "pyqt6-fluent-widgets>=1.11.0",
    "pyqt6-qt6>=6.11.0",
    "pyqt6-sip>=13.11.0",
    "setuptools==70.3.0",
    "sherpa-onnx==1.12.35",
    "webrtcvad==2.0.10",
]
